Lightning

+++ model_development/dataset_management.py
@@ -5,10 +5,10 @@ import yaml
 import getpass
 import numpy as np
 import cv2
+import shutil
 import sys
 sys.path.append("utils")
 from utils.config import Config
-from utils.path_resolver import construct_new_folder
 
 def setup_database_management():
     api_key = getpass.getpass("Enter your Roboflow API key: ")
@@ -33,18 +33,26 @@ def download_dataset(workspace: str, project: str, version: int, annotation_form
     project = rf.workspace(workspace).project(project)
     dataset_version = project.version(version)
 
-    download_location = construct_new_folder(Config.get_development_dataset_path() / f"v{version}data")
-    download_location = construct_new_folder(download_location / Config.get_object_detection_dataset_name())
+    download_location = Config.get_development_dataset_path() / f"v{version}data" / Config.get_object_detection_dataset_name()
+    data_yaml_path = download_location / "data.yaml"
+
+    if download_location.exists():
+        assert data_yaml_path.exists()
+        print("Dataset alraedy downloaded")
+        return data_yaml_path
+
     dataset_version.download(annotation_format, location=str(download_location))
 
-    data_yaml_path = download_location / "data.yaml"
     update_splits_path(data_yaml_path)
 
     return data_yaml_path
 
 def construct_image_classification_folder_structure(object_detection_data_yaml_path: Path):
     object_detection_location = object_detection_data_yaml_path.parent
-    image_classification_location = construct_new_folder(object_detection_location.parent / Config.get_image_classification_dataset_name())
+    image_classification_location = object_detection_location.parent / Config.get_image_classification_dataset_name()
+    if image_classification_location.exists():
+        print("Image classification folder already exists")
+        return image_classification_location, [f for f in image_classification_location.glob("*") if f.is_dir()], True
 
     labels = []
     with open(str(object_detection_data_yaml_path), "r") as f:
@@ -54,7 +62,7 @@ def construct_image_classification_folder_structure(object_detection_data_yaml_p
     for label in labels:
         (image_classification_location / label).mkdir(parents=True)
     
-    return image_classification_location, labels
+    return image_classification_location, labels, False
 
 def extract_boxes(imagefile: Path, labelfile: Path):
     boxes = []
@@ -84,40 +92,52 @@ def extract_boxes(imagefile: Path, labelfile: Path):
     return boxes, label_ids
 
 def construct_image_classification_dataset(object_detection_data_yaml_path: Path) -> Path:
-    image_classification_folder, labels = construct_image_classification_folder_structure(object_detection_data_yaml_path)
+    image_classification_folder, labels, already_constructed = construct_image_classification_folder_structure(object_detection_data_yaml_path)
+    if not already_constructed:
+        object_detection_dataset = object_detection_data_yaml_path.parent
+        for imagefile in object_detection_dataset.rglob("*.jpg"):
+            labelfile = imagefile.parent.parent / "labels" / (imagefile.stem + ".txt")
+            if not labelfile.exists():
+                print(f"Image {imagefile} does not have annotations {labelfile}")
+                continue
+
+            boxes, label_ids = extract_boxes(imagefile, labelfile)
+            assert len(boxes) == len(label_ids), f"Different number of boxes {len(boxes)} and label ids {len(label_ids)}"
+            for i, (box, label_id) in enumerate(zip(boxes, label_ids)):
+                label = labels[label_id]
+                new_imagefile_folder = image_classification_folder / label
+                assert new_imagefile_folder.exists(), f"Folder {new_imagefile_folder} shold have already been created"
+                new_imagefile = imagefile.stem + f"_patch{i}" + imagefile.suffix
+                cv2.imwrite(str(new_imagefile_folder / new_imagefile), box)
+    
     return image_classification_folder
 
-    object_detection_dataset = object_detection_data_yaml_path.parent
-    for imagefile in object_detection_dataset.rglob("*.jpg"):
-        labelfile = imagefile.parent.parent / "labels" / (imagefile.stem + ".txt")
-        if not labelfile.exists():
-            print(f"Image {imagefile} does not have annotations {labelfile}")
-            continue
+def aggregate_all_classes(data_yaml: Path):
+    aggregated_suffix = "_aggregated"
+    object_detection_path = data_yaml.parent
+    aggregated_path = object_detection_path.parent / (object_detection_path.name + aggregated_suffix)
+    aggregated_yaml = aggregated_path / data_yaml.name
+    if aggregated_path.exists():
+        print("Aggregated path already exists")
+        return aggregated_yaml
 
-        boxes, label_ids = extract_boxes(imagefile, labelfile)
-        assert len(boxes) == len(label_ids), f"Different number of boxes {len(boxes)} and label ids {len(label_ids)}"
-        for i, (box, label_id) in enumerate(zip(boxes, label_ids)):
-            label = labels[label_id]
-            new_imagefile_folder = image_classification_folder / label
-            assert new_imagefile_folder.exists(), f"Folder {new_imagefile_folder} shold have already been created"
-            new_imagefile = imagefile.stem + f"_patch{i}" + imagefile.suffix
-            cv2.imwrite(str(new_imagefile_folder / new_imagefile), box)
+    shutil.copytree(object_detection_path, aggregated_path)
 
-def aggregate_all_classes(data_yaml_path: Path):
     aggregated_class = "0"
-    with open(str(data_yaml_path)) as f:
+    with open(str(aggregated_yaml)) as f:
         data_config = yaml.load(f, Loader=yaml.SafeLoader)
     
     data_config["names"] = [Config.get_sharktrack_class()]
+    data_config["path"] = data_config["path"] + aggregated_suffix
     data_config["nc"] = 1
 
-    with open(str(data_yaml_path), "w") as f:
+    with open(str(aggregated_yaml), "w") as f:
         yaml.dump(data_config, f)
     
     splits = ["train", "valid", "test"]
 
     for split in splits:
-        split_path = data_yaml_path.parent / split
+        split_path = aggregated_yaml.parent / split
         for label_file in split_path.rglob("*.txt"):
             print(f"Aggregating classes for file {label_file}")
             new_instances = []
@@ -127,3 +147,5 @@ def aggregate_all_classes(data_yaml_path: Path):
 
             with open(str(label_file), "w") as f:
                 f.writelines(new_instances)
+    
+    return aggregated_yaml
